PET/requirements.txt
rdflib
deepspeed==0.8.0
numpy==1.20
pandas==1.0
scikit-learn==1.2.1
scipy==1.10.0
matplotlib==3.7.0
rapidfuzz==2.13.7
datasets==2.9.0
promptsource==0.2.3
evaluate==0.4.0
pytorch_lightning==1.9.2
tqdm==4.64.1
transformers==4.26.0
tensorboard==2.11.2
wandb==0.13.10
huggingface-hub==0.12.0
sqlparse
nltk